Even more Debug

+++ src/malta.h
@@ -13,7 +13,7 @@ class Malta {
         double integrate(double (*integrand)(double), int n, int n_step, int m);
 
     private:
-        double chi_square(double *integrals, double *errors, int current_step, long double avg_integral);
+        double chi_square(double *integrals, double *errors, int current_step, double avg_integral);
 };
 
 #endif
